requirements.txt
streamlit
google-genai
streamlit-cookies-manager
python-dotenv
